Relearn loops and understanding of function params
After a much needed review, I was able to successfully
impliment the requeired for loop using i to start
at the first index [0], check the i variable to be
valid in regards to the parameter of 'num', and increment
itself after each successful loop by 1 (i++).

Furthermore, using an empty string variable (phrase),
I was able to learn how an accumulator variable can
be useful especially in a scenario such as this.

Moving onto the next exercise.